Cleaned pods and directories
