+++ k8s/templates/mongo/mongo-deployment.yaml
@@ -2,8 +2,6 @@ apiVersion: apps/v1
 kind: Deployment
 metadata: 
   name: mongo
-  labels:
-    chart: mongo
 spec:
   selector:
     matchLabels:
