+++ k8s/templates/mongo/mongo-service.yaml
@@ -7,8 +7,8 @@ metadata:
 spec:
   ports:
   - port: 27017
-    targetPort: 27017
+    nodePort: 30001
     protocol: TCP
-  type: ClusterIP  
+  type: NodePort  
   selector:
-    name: mongo
+    app: mongo
